Add ASV benchmark CI workflow (#139)
* Add ASV workflow

* Apply suggestions from code review

* Forgot to delete folders

* Update asv_bench/benchmarks/combine.py

* Restart with current benchmarks

* Always run benchmark for now.

* Don't test argmax until it's supported in "flox" engine

* Add a skip_slow function.

* try out micromamba

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add os

* try adding asv to  environment file
