--- benchmarks/__init__.py
@@ -1,7 +1,0 @@
-def parameterized(names, params):
-    def decorator(func):
-        func.param_names = names
-        func.params = params
-        return func
-
-    return decorator
