+++ ci/environment.yml
@@ -2,6 +2,7 @@ name: flox-tests
 channels:
   - conda-forge
 dependencies:
+  - asv
   - cachey
   - codecov
   - dask-core
